Auto draw heatmaps created
Script to draw heatmaps from file in a directory whose filename matches a pattern has been created. Main needs to be updated to accept command line arguments and directories are hard coded for now. Will only run on python 3.9 (not 3.10) and does not work in vscode due to matplotlib library compatibility issues