Group transitions from an AOI in the same file and use AOI name in headers
Code originally grouped the transitions based on the AOI the participant transitioned to (i.e., the second AOI in the pair). Now the transitions are grouped based on the first AOI in the pair. Additionally, the csv files use the actual AOI name rather a default aoi in the headers. The combined file still uses aoi because there are multiple AOIs